re coded quiz functions to new database specs
